Files refactor
