--- modules/ipfs-cpinner-client/src/errors.ts
@@ -1,3 +1,0 @@
-export const NO_DID = 'No DID'
-export const NO_SIGNER = 'No Signer'
-export const NO_SERVICE_DID = 'No service DID'
