+++ modules/ipfs-cpinner-client/test/swap-content.test.ts
@@ -1,4 +1,4 @@
-import DataVaultWebClient, { Signer } from '../src'
+import DataVaultWebClient from '../src'
 import { startService, identityFactory, deleteDatabase, testTimestamp } from './util'
 import { Server } from 'http'
 import { Connection } from 'typeorm'
@@ -6,6 +6,7 @@ import { IpfsPinnerProvider } from '@rsksmart/ipfs-cpinner-provider'
 import MockDate from 'mockdate'
 import ipfsHash from 'ipfs-only-hash'
 import LocalStorageMockFactory from './localStorageMockFactory'
+import { Signer } from '../src/types'
 
 jest.setTimeout(10000)
